backend/requirements.txt
praw>=7.7.0
fastapi>=0.104.0
uvicorn>=0.24.0
google-genai>=1.0.0
python-dotenv>=1.0.0
httpx>=0.25.0
aiosqlite>=0.19.0
